+++ sdmx/source/estat3.py
@@ -12,8 +12,7 @@ class Source(ESTAT):
         """
         super().modify_request_args(kwargs)
 
-        # NB this is an indirect test for resource_type != 'data'; because of
-        #    the way the hook is called, resource_type is not available
-        #    directly.
+        # NB this is an indirect test for resource_type != 'data'; because of the way
+        #    the hook is called, resource_type is not available directly.
         if "key" not in kwargs:
-            kwargs.setdefault("provider", "ESTAT")
+            kwargs.setdefault("agency_id", "ESTAT")